[DESIGN]: Workshop
Update the backpack when a recipe is selected to display only the needed ingedients.
Once all the needed nb of ingredients for the recipe have been transfered, the "Craft" button will appear.
After clicking the "Craft" button, the Craft button will disappear and the newly crafted item will be added to the backpack. The backpack will be displayed entirely, the ingredient table will be set blank as well as the image designing the selected recipe.